power: Silence charger log spam
Charger log spam is so invasive that Google employees have to ask bug
reporters to collect fresh logs that are a few hours removed from when a
charger is connected. The charger log spam not only comes from the
kernel, but also healthd in userspace. To that effect, remove
userspace's privilege of writing to /dev/kmsg. And put a damper on the
kernel-side log spam as well.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Kazuki Hashimoto <kaz205@tuta.io>