arm64: dts: taimen: Reduce panel-on delay
This reduces the total delay from 170 ms to 15 ms.

In commit 11fc88d1339b27ebb6009a3d8ce54cd8bac3b28a, Google and LG
described issues with flickering when turning the display on in
low-temperature environments. LG proposed two independent solutions for
the problem:
    - Increasing delay
    - Increasing interval width

Google applied the former solution as a quick fix, but decided to apply
the latter fix in a later update. This means that the increased delay
can be reverted in favor of only keeping the interval width fix, which
is still preserved in the panel-on commands, so there shouldn't be any
flickering issues as a result of this.

Test: display turns on and colors are correct
Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: Kazuki Hashimoto <kaz205@tuta.io>